Create worker update
First you run TimeServer.Java in which a thread is intitially created (WorkerCountThread.Java) that will create a worker (WorkerThread.Java) as needed. WorkerCountThread.Java runs a while loop checking the public static variable in TimeServer.Java and creates a worker (WorkerThread.Java) if there aren't any running and will create another worker when the size of the list of requests increases past 2 (Currently limited so it only creates a max of two workers). WorkerThread.Java is where the string/files will need to be filtered. 

WorkerCountThread.Java will create a socket and send "WORKER" through the socket as UTF data to the server (TimeServer.Java). This will allow TimeServer.Java to know that the socket it found is for a WorkerThread which is what it will create.

Next you run SimpleClient.Java which will create a ClientThread.Java which will handle input and output to the server. SimpleClient.Java also reads input from the computer and outputs it to ClientThread.Java. SimplleClient.Java will create a socket and output ("CLIENT") as UTF data so that when TimeServer.Java detects the socket it will know that it needs to create a TimeClient.Java thread in which all the user input is processed to allow for making a request, checking status, stopping etc...